requirements.txt
aniso8601==9.0.1
antiorm==1.2.1
attrs==23.1.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
clickclick==20.10.2
connexion==2.14.2
db==0.1.1
dnspython==2.3.0
et-xmlfile==1.1.0
Flask==2.2.5
Flask-Cors==4.0.0
Flask-Mail==0.9.1
Flask-RESTful==0.3.10
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.7.0
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
jsonschema-specifications==0.1.3
lxml==4.9.3
MarkupSafe==2.1.3
multidict==6.0.4
openpyxl==3.1.2
packaging==23.1
pymongo==4.4.1
pyrsistent==0.19.3
python-docx==0.8.11
python-dotenv==0.21.1
pytz==2023.3
PyYAML==6.0.1
referencing==0.8.9
requests==2.31.0
# rpds-py==0.9.2
six==1.16.0
swagger-ui-bundle==0.0.9
typing_extensions==4.7.1
urllib3==1.26.15
Werkzeug==2.2.3
yarl==1.9.2
zipp==3.15.0
